sampling_parameter/requirements.txt
accelerate==1.10.1
agno==2.0.4
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asttokens==3.0.0
# Editable install with no version control (athena==0.1)
-e c:\users\praty\desktop\shortnote\resume projects\athena
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
boto3==1.40.27
botocore==1.40.27
Bottleneck==1.6.0
Brotli==1.1.0
build==1.3.0
CacheControl==0.14.3
cached_path==1.7.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cleo==2.1.0
click==8.2.1
colorama==0.4.6
colour==0.1.5
comm==0.2.3
contourpy==1.3.2
crashtest==0.4.1
cssselect2==0.8.0
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.6.7
datasets==4.0.0
debugpy==1.8.16
decorator==5.2.1
deep-translator==1.11.4
dill==0.3.8
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.17.0
dulwich==0.22.8
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.7
email_validator==2.2.0
encodec==0.1.1
exa-py==1.16.1
executing==2.2.0
f5_tts @ git+https://github.com/ai4bharat/IndicF5.git@b334bb3b1f6a027420cd02c8ffb1fc95b146c6e8
fastapi==0.116.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.2
fastjsonschema==2.21.2
filelock==3.19.1
filetype==1.2.0
findpython==0.6.3
fonttools==4.58.5
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
glcontext==3.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.113.0
google-cloud-bigquery==3.37.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.36.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googletrans==4.0.2
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
installer==0.7.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isosurfaces==0.1.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==25.6.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.79
langchain-exa==1.0.0
langchain-google-genai==2.1.10
langchain-google-vertexai==2.1.0
langchain-tavily==0.2.12
langchain-text-splitters==0.3.11
langdetect-py==1.1.1
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.27
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.2
# Editable install with no version control (manimgl==1.7.1)
-e c:\users\praty\desktop\manim-master\manim-master
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numexpr==2.12.1
numpy==2.3.3
omegaconf==2.3.0
openai==1.107.2
opencv-python==4.11.0.86
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
parso==0.8.4
pathlib==1.0.1
pbs-installer==2025.8.28
pillow==11.3.0
pip-chill==1.0.3
pipreqs==0.4.13
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.4
poetry-core==2.1.3
polars==1.34.0
polars-runtime-32==1.34.0
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pydyf==0.11.0
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.2
PyMuPDF==1.26.4
pymupdf4llm==0.0.27
PyOpenGL==3.1.9
pyparsing==3.2.3
pyperclip==1.9.0
pyphen==0.17.2
pypinyin==0.55.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.0.2
RapidFuzz==3.14.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.8
rignore==0.5.1
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
screeninfo==0.8.1
sentry-sdk==2.32.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
skia-pathops==0.8.0.post2
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.46.2
streamlit-drawable-canvas-fix==0.9.7
svgelements==1.9.6
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.9.9
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.8.0
torchdiffeq==0.2.5
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
transformers-stream-generator==0.0.5
trove-classifiers==2025.8.26.11
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.204
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
virtualenv==20.27.0
vocos==0.1.0
wandb==0.21.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasyprint==66.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
wheel==0.44.0
win32_setctime==1.2.0
wsproto==1.2.0
WTForms==3.1.2
x-transformers==2.7.4
xarray==2025.6.1
xesmf==0.8.10
xxhash==3.5.0
xyzservices==2025.4.0
yarg==0.1.10
yarl==1.20.1
zict==3.0.0
zipp==3.23.0
zopfli==0.2.3.post1
zstandard==0.25.0
